GETP-199 rename: member 컴포넌트의 exception 모듈을 응집성이 높은 방향으로 분리
