+++ src/main/java/es/princip/getp/domain/member/service/FailedToSaveProfileImageException.java
@@ -1,4 +1,4 @@
-package es.princip.getp.domain.member.exception;
+package es.princip.getp.domain.member.service;
 
 import es.princip.getp.common.exception.ApiErrorException;
 import es.princip.getp.common.exception.ErrorDescription;
